# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "MonColis Particulier API",
        "version": "0.1.0",
        "description": "API backend de **MonColis Particulier** (app mobile Flutter).\r\n\r\n- Base : `/api` \u2014 JSON en camelCase, r\u00e9ponses envelopp\u00e9es dans `{ \"data\": ... }`.\r\n- Authentification : `POST /login_check` (ou `/register`) renvoie un `token` \u00e0 passer en `Authorization: Bearer {token}`.\r\n- En-t\u00eate `client-app-code` : obligatoire lorsque le serveur a des codes configur\u00e9s.\r\n- Erreurs m\u00e9tier : `{ \"message\": \"...\", \"code\": \"INVALID_CREDENTIALS\" }` \u2014 le `code` est stable, l'app peut s'y fier.\r\n- Erreurs de validation : `422 { \"message\": \"...\", \"errors\": { \"champ\": [\"...\"] } }`.\r\n\r\nSwagger UI : [/docs/swagger](/docs/swagger) \u2014 Document OpenAPI brut : [/docs/api.json](/docs/api.json)."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
...
    "security": [
        {
            "bearerAuth": [],
            "clientAppCode": []
        }
    ],
    "paths": {
        "/auth/otp/request": {
            "post": {
                "operationId": "auth.otp.request",
                "description": "Envoie un code \u00e0 6 chiffres par SMS. M\u00eame r\u00e9ponse que le num\u00e9ro ait d\u00e9j\u00e0 un compte ou non.\n`devCode` n'est renvoy\u00e9 qu'en environnement local.\n\nErreur m\u00e9tier : `429 OTP_COOLDOWN` (nouvel envoi avant `resendIn` secondes).",
                "summary": "Recevoir un code de connexion",
                "tags": [
                    "Connexion par SMS"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhoneCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "const": "sms"
                                                },
                                                "destination": {
                                                    "type": "string"
                                                },
                                                "expiresIn": {
                                                    "type": "integer"
                                                },
                                                "resendIn": {
                                                    "type": "integer"
                                                },
                                                "devCode": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "channel",
                                                "destination",
                                                "expiresIn",
                                                "resendIn"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": [
                    {
                        "clientAppCode": []
                    }
                ]
            }
        },
...
        "/profile": {
            "get": {
                "operationId": "profile.show",
                "description": "Inclut `completion` (pourcentage + champs manquants parmi `firstName`, `lastName`,\n`phoneVerified`, `email`, `avatar`) pour le bandeau \u00ab Compl\u00e9tez votre profil \u00bb.",
                "summary": "Mon profil",
                "tags": [
                    "Profil"
                ],
                "responses": {
                    "200": {
                        "description": "`ProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProfileResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "preferences"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "profile.update",
                "description": "Mise \u00e0 jour partielle : seuls les champs envoy\u00e9s sont modifi\u00e9s. Changer l'e-mail remet\n`emailVerified` \u00e0 `false`. Le num\u00e9ro ne se change pas ici (voir \u00ab S\u00e9curit\u00e9 du compte \u00bb).",
                "summary": "Modifier mon profil",
                "tags": [
                    "Profil"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProfileResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "preferences"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "profile.destroy",
                "description": "Efface les donn\u00e9es personnelles, r\u00e9voque tous les jetons et lib\u00e8re le num\u00e9ro / l'e-mail.\n`password` est demand\u00e9 si le compte en a un ; sinon (compte cr\u00e9\u00e9 par code SMS), `confirm: true`.\nL'historique des livraisons est conserv\u00e9 de fa\u00e7on anonyme.",
                "summary": "Supprimer mon compte",
                "tags": [
                    "Profil"
                ],
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "confirm",
                        "in": "query",
                        "required": true,
                        "description": "Account without password (created by SMS code): explicit confirmation instead.",
                        "schema": {
                            "enum": [
                                "yes",
                                "on",
                                "1",
                                1,
                                "true",
                                true
                            ]
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 500
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Jeton renvoy\u00e9 par /login_check ou /register.",
                "scheme": "bearer"
            },
            "clientAppCode": {
                "type": "apiKey",
                "description": "Code de l'application cliente (v\u00e9rifi\u00e9 si MONCOLIS_CLIENT_APP_CODES est renseign\u00e9).",
                "in": "header",
                "name": "client-app-code"
            }
        },
        "schemas": {
...
            "Gender": {
                "type": "string",
                "enum": [
                    "male",
                    "female",
                    "other"
                ],
                "title": "Gender"
            },
...
            "PhoneCodeRequest": {
                "type": "object",
                "description": "Step 1 of the SMS sign-in flow: the phone number.",
                "properties": {
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "PhoneCodeRequest"
            },
            "PreferencesResource": {
                "type": "object",
                "properties": {
                    "language": {
                        "type": "string"
                    },
                    "theme": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "notifyPush": {
                        "type": "boolean"
                    },
                    "notifySms": {
                        "type": "boolean"
                    },
                    "notifyEmail": {
                        "type": "boolean"
                    },
                    "marketingOptIn": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "language",
                    "theme",
                    "notifyPush",
                    "notifySms",
                    "notifyEmail",
                    "marketingOptIn"
                ],
                "title": "PreferencesResource"
            },
            "ProfileResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "firstName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "lastName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fullName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "phoneVerified": {
                        "type": "boolean"
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "emailVerified": {
                        "type": "boolean"
                    },
                    "hasPassword": {
                        "type": "boolean",
                        "description": "false for accounts created by SMS code until a password is set."
                    },
                    "avatarUrl": {
                        "type": "null"
                    },
                    "gender": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "birthDate": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "countryCode": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "profile": {
                        "type": "string",
                        "description": "Legacy field read by the Flutter UserModel.",
                        "const": "PARTICULIER"
                    },
                    "memberSince": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "completion": {
                        "type": "object",
                        "properties": {
                            "percent": {
                                "type": "integer"
                            },
                            "missing": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "required": [
                            "percent",
                            "missing"
                        ]
                    },
                    "preferences": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PreferencesResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "required": [
                    "id",
                    "username",
                    "firstName",
                    "lastName",
                    "fullName",
                    "displayName",
                    "phone",
                    "phoneVerified",
                    "email",
                    "emailVerified",
                    "hasPassword",
                    "avatarUrl",
                    "gender",
                    "birthDate",
                    "countryCode",
                    "profile",
                    "memberSince",
                    "completion"
                ],
                "title": "ProfileResource"
            },
...
            "UpdateProfileRequest": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 80
                    },
                    "lastName": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 80
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "minLength": 3,
                        "maxLength": 60
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "email",
                        "maxLength": 255
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Gender"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "birthDate": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date"
                    },
                    "countryCode": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "minLength": 2,
                        "maxLength": 2
                    },
                    "phone": {
                        "type": "string",
                        "description": "The phone number is the account identity: it only changes through the SMS code flow."
                    }
                },
                "title": "UpdateProfileRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}